refactoring key generation info
