+++ ts/features/lollipop/hooks/useLollipopLoginSource.tsx
@@ -1,11 +1,16 @@
+import { generate, deleteKey } from "@pagopa/io-react-native-crypto";
 import { pipe } from "fp-ts/lib/function";
 import * as O from "fp-ts/lib/Option";
 import * as TE from "fp-ts/lib/TaskEither";
 import { useState } from "react";
 import { useIODispatch, useIOSelector } from "../../../store/hooks";
 import { isLollipopEnabledSelector } from "../../../store/reducers/backendStatus";
-import { trackLollipopIdpLoginFailure } from "../../../utils/analytics";
-import { taskRegenerateKey } from "../../../utils/crypto";
+import {
+  trackLollipopIdpLoginFailure,
+  trackLollipopKeyGenerationFailure,
+  trackLollipopKeyGenerationSuccess
+} from "../../../utils/analytics";
+import { toCryptoError } from "../utils/crypto";
 import { useOnFirstRender } from "../../../utils/hooks/useOnFirstRender";
 import {
   lollipopRemovePublicKey,
@@ -14,6 +19,12 @@ import {
 import { lollipopKeyTagSelector } from "../store/reducers/lollipop";
 import { LoginSourceAsync } from "../types/LollipopLoginSource";
 import { DEFAULT_LOLLIPOP_HASH_ALGORITHM_SERVER } from "../utils/login";
+
+const taskRegenerateKey = (keyTag: string) =>
+  pipe(
+    TE.tryCatch(() => deleteKey(keyTag), toCryptoError),
+    TE.chain(() => TE.tryCatch(() => generate(keyTag), toCryptoError))
+  );
 
 export const useLollipopLoginSource = (loginUri?: string) => {
   const [loginSource, setLoginSource] = useState<LoginSourceAsync>({
@@ -82,9 +93,11 @@ export const useLollipopLoginSource = (loginUri?: string) => {
           publicKey: O.some(key)
         });
         dispatch(lollipopSetPublicKey({ publicKey: key }));
+        trackLollipopKeyGenerationSuccess(key.kty);
       }),
       TE.mapLeft(error => {
         trackLollipopIdpLoginFailure(error.message);
+        trackLollipopKeyGenerationFailure(error.message);
         setDeprecatedLoginUri(loginUri);
         dispatch(lollipopRemovePublicKey());
       })
