+++ ts/features/lollipop/httpSignature/__tests__/signature.test.ts
@@ -8,7 +8,7 @@ import {
   customContentSignatureBases,
   CustomContentBaseSignature
 } from "../../utils/fetch";
-import { KeyInfo } from "../../../../utils/crypto";
+import { KeyInfo } from "../../utils/crypto";
 import { constants } from "../constants";
 import {
   generateSignatureInput,
