+++ ts/features/lollipop/saga/index.ts
@@ -1,16 +1,28 @@
 import * as O from "fp-ts/lib/Option";
+import { pipe } from "fp-ts/lib/function";
+import * as T from "fp-ts/lib/Task";
+import * as TE from "fp-ts/lib/TaskEither";
 import { v4 as uuid } from "uuid";
 import { put, select, call } from "typed-redux-saga/macro";
-import { getPublicKey } from "@pagopa/io-react-native-crypto";
+import {
+  deleteKey,
+  generate,
+  getPublicKey,
+  PublicKey
+} from "@pagopa/io-react-native-crypto";
+import { jwkThumbprintByEncoding } from "jwk-thumbprint";
 import { lollipopKeyTagSelector } from "../store/reducers/lollipop";
 import {
   lollipopKeyTagSave,
+  lollipopRemovePublicKey,
   lollipopSetPublicKey
 } from "../store/actions/lollipop";
+import { KeyInfo, toCryptoError } from "../utils/crypto";
 import {
-  cryptoKeyGenerationSaga,
-  deletePreviousCryptoKeyPair
-} from "../../../sagas/startup/generateCryptoKeyPair";
+  trackLollipopKeyGenerationFailure,
+  trackLollipopKeyGenerationSuccess
+} from "../../../utils/analytics";
+import { DEFAULT_LOLLIPOP_HASH_ALGORITHM_CLIENT } from "../utils/login";
 
 export function* generateLollipopKeySaga() {
   const maybeOldKeyTag = yield* select(lollipopKeyTagSelector);
@@ -28,6 +40,7 @@ export function* generateLollipopKeySaga() {
       // a public key tied with it.
       const publicKey = yield* call(getPublicKey, maybeOldKeyTag.value);
       yield* put(lollipopSetPublicKey({ publicKey }));
+      yield* call(trackLollipopKeyGenerationSuccess, publicKey.kty);
     } catch {
       // If there is no key it could be for two reasons:
       // - The user have a recent app and they logged out (the key is deleted).
@@ -47,3 +60,104 @@ export function* deleteCurrentLollipopKeyAndGenerateNewKeyTag() {
   const newKeyTag = uuid();
   yield* put(lollipopKeyTagSave({ keyTag: newKeyTag }));
 }
+
+/**
+ * Generates a new crypto key pair.
+ */
+function* cryptoKeyGenerationSaga(
+  keyTag: string,
+  previousKeyTag: O.Option<string>
+) {
+  // Every new login we need to regenerate a brand new key pair.
+  yield* call(deletePreviousCryptoKeyPair, previousKeyTag);
+  yield* call(generateCryptoKeyPair, keyTag);
+}
+
+/**
+ * Deletes a previous saved crypto key pair.
+ */
+function* deletePreviousCryptoKeyPair(keyTag: O.Option<string>) {
+  if (O.isSome(keyTag)) {
+    yield* call(deleteCryptoKeyPair, keyTag.value);
+  }
+}
+
+/**
+ * Deletes the crypto key pair corresponding to the provided `keyTag`.
+ */
+function* deleteCryptoKeyPair(keyTag: string) {
+  // Key is persisted even after uninstalling the application on iOS.
+  const keyAlreadyExistsOnKeystore = yield* call(checkPublicKeyExists, keyTag);
+
+  if (keyAlreadyExistsOnKeystore) {
+    try {
+      yield* call(deleteKey, keyTag);
+      yield* put(lollipopRemovePublicKey());
+    } catch (e) {
+      const { message } = toCryptoError(e);
+      yield* call(trackLollipopKeyGenerationFailure, message);
+    }
+  }
+}
+
+const checkPublicKeyExists = (keyTag: string) =>
+  pipe(
+    TE.tryCatch(
+      () => getPublicKey(keyTag),
+      () => false
+    ),
+    TE.map(_ => true),
+    TE.getOrElse(() => T.of(false))
+  )();
+
+/**
+ * Generates a new crypto key pair.
+ */
+function* generateCryptoKeyPair(keyTag: string) {
+  try {
+    // Remove an already existing key with the same tag.
+    yield* call(deleteCryptoKeyPair, keyTag);
+
+    const publicKey = yield* call(generate, keyTag);
+    yield* put(lollipopSetPublicKey({ publicKey }));
+    yield* call(trackLollipopKeyGenerationSuccess, publicKey.kty);
+  } catch (e) {
+    const { message } = toCryptoError(e);
+    yield* call(trackLollipopKeyGenerationFailure, message);
+  }
+}
+
+export function generateKeyInfo(
+  maybeKeyTag: O.Option<string>,
+  maybePublicKey: O.Option<PublicKey>
+) {
+  return pipe(
+    maybeKeyTag,
+    O.chain(keyTag =>
+      pipe(
+        maybePublicKey,
+        O.map(publicKey => keyInfoFromKeyTagAndPublicKey(keyTag, publicKey))
+      )
+    ),
+    O.getOrElse(defaultKeyInfo)
+  );
+}
+
+const keyInfoFromKeyTagAndPublicKey = (
+  keyTag: string,
+  publicKey: PublicKey
+): KeyInfo => ({
+  keyTag,
+  publicKey,
+  publicKeyThumbprint: jwkThumbprintByEncoding(
+    publicKey,
+    DEFAULT_LOLLIPOP_HASH_ALGORITHM_CLIENT,
+    "base64url"
+  )
+});
+
+const defaultKeyInfo = (): KeyInfo => ({
+  keyTag: undefined,
+  publicKey: undefined,
+  publicKeyThumbprint: undefined
+});
