+++ ts/sagas/startup.ts
@@ -67,7 +67,10 @@ import {
   lollipopKeyTagSelector,
   lollipopPublicKeySelector
 } from "../features/lollipop/store/reducers/lollipop";
-import { generateLollipopKeySaga } from "../features/lollipop/saga";
+import {
+  generateKeyInfo,
+  generateLollipopKeySaga
+} from "../features/lollipop/saga";
 import { IdentificationResult } from "../store/reducers/identification";
 import { pendingMessageStateSelector } from "../store/reducers/notifications/pendingMessage";
 import {
@@ -143,10 +146,6 @@ import { completeOnboardingSaga } from "./startup/completeOnboardingSaga";
 import { watchLoadMessageById } from "./messages/watchLoadMessageById";
 import { watchThirdPartyMessageSaga } from "./messages/watchThirdPartyMessageSaga";
 import { checkNotificationsPreferencesSaga } from "./startup/checkNotificationsPreferencesSaga";
-import {
-  generateKeyInfo,
-  trackMixpanelCryptoKeyPairEvents
-} from "./startup/generateCryptoKeyPair";
 
 const WAIT_INITIALIZE_SAGA = 5000 as Millisecond;
 const navigatorPollingTime = 125 as Millisecond;
@@ -208,11 +207,7 @@ export function* initializeApplicationSaga(): Generator<
   // user profile.
   yield* put(resetProfileState());
 
-  // Generate key for lollipop
-  // TODO Once the lollipop feature is spread to the all the user base,
-  // consider refactoring even more by removing this, when
-  // https://pagopa.atlassian.net/browse/LLK-38 has been fixed.
-  // For now we need to generate a key in the application startup flow
+  // We need to generate a key in the application startup flow
   // to use this information on old app version already logged in users.
   // Here we are blocking the application startup, but we have the
   // the profile loading spinner active.
@@ -363,11 +358,6 @@ export function* initializeApplicationSaga(): Generator<
 
   // check if the user expressed preference about mixpanel, if not ask for it
   yield* call(askMixpanelOptIn);
-
-  // Track crypto key generation info
-  if (O.isSome(keyTag)) {
-    yield* call(trackMixpanelCryptoKeyPairEvents, keyTag.value);
-  }
 
   if (hasPreviousSessionAndPin) {
     // We have to retrieve the pin here and not on the previous if-condition (same guard)
