+++ ts/utils/analytics.ts
@@ -6,7 +6,6 @@ import { mixpanelTrack } from "../mixpanel";
 import { ReminderStatusEnum } from "../../definitions/backend/ReminderStatus";
 import { UIMessageId } from "../store/reducers/entities/messages/types";
 import { ServiceId } from "../../definitions/backend/ServiceId";
-import { KeyGenerationInfo } from "./crypto";
 
 const blackListRoutes: ReadonlyArray<string> = [];
 
@@ -140,15 +139,15 @@ export function trackThirdPartyMessageAttachmentUserAction(
 // End of premium events
 
 // Lollipop events
-export function trackLollipopKeyGenerationSuccess(keyInfo: KeyGenerationInfo) {
+export function trackLollipopKeyGenerationSuccess(keyType?: string) {
   void mixpanelTrack("LOLLIPOP_KEY_GENERATION_SUCCESS", {
-    kty: keyInfo.keyType
+    kty: keyType
   });
 }
 
-export function trackLollipopKeyGenerationFailure(keyInfo: KeyGenerationInfo) {
+export function trackLollipopKeyGenerationFailure(reason: string) {
   void mixpanelTrack("LOLLIPOP_KEY_GENERATION_FAILURE", {
-    reason: keyInfo.errorCode
+    reason
   });
 }
